% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areas_Halibut_sf.R
\docType{data}
\name{Areas_Halibut_sf}
\alias{Areas_Halibut_sf}
\title{Areas_Halibut_sf Metadata}
\format{
sf polygon
\describe{
\item{Strata}{polygon identifier}
}
}
\usage{
Areas_Halibut_sf
}
\description{
These are the Halibut areas converted to an sf object. A
This data was provided by the halibut unit on Nov 18, 2020
in pbsmapping format.
}
\keyword{datasets}
