% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areas_Scallop_sf.R
\docType{data}
\name{Areas_Scallop_sf}
\alias{Areas_Scallop_sf}
\title{Areas_Scallop_sf Metadata}
\format{
sf polygon
\describe{
\item{AREA}{polygon identifier}
\item{Type}{one of "Inshore", "Offshore" or "Survey"}
\item{Label}{Original Polygon Identifier}
}
}
\usage{
Areas_Scallop_sf
}
\description{
These are the Scallop areas converted to an sf object. All Offshore
Polygon StrataID values have had 1000 added to them to avoid duplication of
StrataID values with the Inshore data.
}
\keyword{datasets}
