% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areas_Shrimp_sf.R
\docType{data}
\name{Areas_Shrimp_sf}
\alias{Areas_Shrimp_sf}
\title{Areas_Shrimp_sf Metadata}
\format{
sf polygon
\describe{
\item{POLY_ID}{polygon identifier}
\item{BOX_NAME}{polygon Name}
}
}
\usage{
Areas_Shrimp_sf
}
\description{
These are the shrimp areas converted to an sf object.
}
\keyword{datasets}
