% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areas_Snowcrab_sf.R
\docType{data}
\name{Areas_Snowcrab_sf}
\alias{Areas_Snowcrab_sf}
\title{Areas_Snowcrab_sf Metadata}
\format{
sf polygon
\describe{
\item{AREA1}{The most general areas - cfa4x, cfanorth and cfasouth}
\item{AREA2}{More detailed aread, sunch as cfa20, cfa21, etc}
\item{AREA3}{Most deailed areas, such as cfa20inner, cfa20outer, etc}
}
}
\usage{
Areas_Snowcrab_sf
}
\description{
These are the Snowcrab areas converted to an sf object.
This data was provided by the snowcrab unit on Nov 19, 2020 as columns of
latitudes and longitudes.
}
\keyword{datasets}
