% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areas_Snowcrab_Slope_sf.R
\docType{data}
\name{Areas_Snowcrab_Slope_sf}
\alias{Areas_Snowcrab_Slope_sf}
\title{Areas_Snowcrab_Slope_sf Metadata}
\format{
sf polygon
\describe{
\item{AREA1}{The most general slope area - cfa}
\item{AREA2}{More detailed slope areas, such as cfa23slope, cfa24slope, etc}
}
}
\usage{
Areas_Snowcrab_Slope_sf
}
\description{
These are the Snowcrab slope areas converted to an sf object.
This data was provided by the snowcrab unit on Nov 19, 2020 as columns of
latitudes and longitudes.  The slope areas will be made available as a
separate object.
}
\keyword{datasets}
