% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Areas_Surfclam_sf.R
\docType{data}
\name{Areas_Surfclam_sf}
\alias{Areas_Surfclam_sf}
\title{Areas_Surfclam_sf Metadata}
\format{
sf polygon
\describe{
\item{AREA}{polygon identifier}
}
}
\usage{
Areas_Surfclam_sf
}
\description{
These are the clam Assessment areas converted to an sf object.
This data was provided by the surfclam unit on Nov 17, 2020
in pbsmapping format.
}
\keyword{datasets}
