% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClamAreas_sf.R
\docType{data}
\name{ClamAreas_sf}
\alias{ClamAreas_sf}
\title{ClamAreas_sf Metadata}
\format{
sf polygon
\describe{
\item{AREA}{polygon identifier}
}
}
\usage{
ClamAreas_sf
}
\description{
These are the clam fishing areas converted to an sf object.
This data was provided by the surfclam unit on Nov 17, 2020
in pbsmapping format.
}
\keyword{datasets}
