% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeorgesBankDiscardZones_sf.R
\docType{data}
\name{GeorgesBankDiscardZones_sf}
\alias{GeorgesBankDiscardZones_sf}
\title{GeorgesBankDiscardZones_sf Metadata}
\format{
sf polygon
\describe{
\item{Id}{ESRI ArcMap Field - a number corresponding to the letter code in the Zone field (e.g. 1 for "Zone A", 2 for "Zone B", etc.)}
\item{Zone}{Discard Zone - the word "Zone" plus an alphabetical identifier (e.g. "Zone A")}
}
}
\source{
https://science-catalogue.canada.ca/record=4039145~S6
}
\usage{
GeorgesBankDiscardZones_sf
}
\description{
These are the discard zones for the Georges Bank groundfish fishery.
They were defined based on fishing concentration and homogeneity of species
composition in Gavaris, Van Eeckhaute, and Clark (2007).
}
\keyword{datasets}
