% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2Min.R
\docType{data}
\name{grid2Min}
\alias{grid2Min}
\title{grid2Min Metadata}
\format{
SpatialPolygonsDataFrame frame with 1 data column
\describe{
\item{PUID}{polygon identifier}
}
}
\usage{
grid2Min
}
\description{
This is a mesh of square 2 minute x 2 minute grid cells.  The northern 
boundary is  48 12'30"N and the eastern boundary is 56 0'30"W.  The southern 
and western boundaries follow the Canadian EEZ.  No cells cover the French 
"Keyhole"
}
\details{
if{html}{figure{grid2Min.png}{options:alt="Figure: grid2Min coverage area"}}
}
\keyword{datasets}
