% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2Min_sf.R
\docType{data}
\name{grid2Min_sf}
\alias{grid2Min_sf}
\title{grid2Min_sf Metadata}
\format{
sf polygon
\describe{
\item{PUID}{polygon identifier}
}
}
\usage{
grid2Min_sf
}
\description{
This is a mesh of square 2 minute x 2 minute grid cells.  The northern
boundary is  48 12'30"N and the eastern boundary is 56 0'30"W.  The southern
and western boundaries follow the Canadian EEZ.  No cells cover the French
"Keyhole"
}
\keyword{datasets}
