% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grids_Lobster_sf.R
\docType{data}
\name{Grids_Lobster_sf}
\alias{Grids_Lobster_sf}
\title{Grids_Lobster_sf Metadata}
\format{
sf polygon
\describe{
\item{GRID}{polygon identifier}
}
}
\usage{
Grids_Lobster_sf
}
\description{
These are the lobster grids converted to an sf object.
This data was provided by the lobster unit on Nov 17, 2020
in pbsmapping format.
}
\keyword{datasets}
