% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HalibutStrata_sf.R
\docType{data}
\name{HalibutStrata_sf}
\alias{HalibutStrata_sf}
\title{HalibutStrata_sf Metadata}
\format{
sf polygon
\describe{
\item{Strata}{polygon identifier}
}
}
\usage{
HalibutStrata_sf
}
\description{
These are the Halibut Strata converted to an sf object. A
This data was provided by the halibut unit on Nov 18, 2020
in pbsmapping format.
}
\keyword{datasets}
