% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFAs_sf.R
\docType{data}
\name{LFAs_sf}
\alias{LFAs_sf}
\title{LFAs_sf Metadata}
\format{
sf polygon
\describe{
\item{LFA}{Lobster Fishing Area}
}
}
\usage{
LFAs_sf
}
\description{
These are the lobster fishing areas converted to an sf object.
This data was provided by the lobster unit on Nov 17, 2020
in pbsmapping format.
}
\keyword{datasets}
