% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LobsterGrids_sf.R
\docType{data}
\name{LobsterGrids_sf}
\alias{LobsterGrids_sf}
\title{LobsterGrids_sf Metadata}
\format{
sf polygon
\describe{
\item{GRID}{polygon identifier}
}
}
\usage{
LobsterGrids_sf
}
\description{
These are the lobster grids converted to an sf object.
This data was provided by the lobster unit on Nov 17, 2020
in pbsmapping format.
}
\keyword{datasets}
