% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarStrata4VSW_sf.R
\docType{data}
\name{MarStrata4VSW_sf}
\alias{MarStrata4VSW_sf}
\title{MarStrata4VSW_sf Metadata}
\format{
sf polygon
\describe{
\item{StrataID}{polygon identifier}
}
}
\usage{
MarStrata4VSW_sf
}
\description{
These are the polygons from "<PED shared drive>\\Spatial\\Science\\Strata\\Maritimes\\PED_4VsW.shp"
- a Maritimes 4VSW strata layer, converted to an sf object.
}
\keyword{datasets}
