% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarStrata_sf.R
\docType{data}
\name{MarStrata_sf}
\alias{MarStrata_sf}
\title{MarStrata_sf Metadata}
\format{
sf multipolygon
\describe{
\item{StrataID}{polygon identifier}
}
}
\usage{
MarStrata_sf
}
\description{
These are the polygons from Maritimes strata, converted to an sf object from
the original "MaritimesRegionEcosystemAssessmentStrata(2014-).shp" shapefile.
}
\keyword{datasets}
