% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScallopFAs_sf.R
\docType{data}
\name{ScallopFAs_sf}
\alias{ScallopFAs_sf}
\title{ScallopFAs_sf Metadata}
\format{
sf polygon
\describe{
\item{StrataID}{polygon identifier}
\item{Type}{one of "Inshore", "Offshore" or "Survey"}
\item{Label}{Original Polygon Identifier}
}
}
\usage{
ScallopFAs_sf
}
\description{
These are the Scallop fishing areas converted to an sf object. All Offshore
Polygon StrataID values have had 1000 added to them to avoid duplication of
StrataID values with the Inshore data.
}
\keyword{datasets}
