% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShrimpFAs_sf.R
\docType{data}
\name{ShrimpFAs_sf}
\alias{ShrimpFAs_sf}
\title{ShrimpFAs_sf Metadata}
\format{
sf polygon
\describe{
\item{POLY_ID}{polygon identifier}
\item{BOX_NAME}{polygon Name}
}
}
\usage{
ShrimpFAs_sf
}
\description{
These are the shrimp fishing areas converted to an sf object.
}
\keyword{datasets}
