% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPAs_Scallop_sf.R
\docType{data}
\name{SPAs_Scallop_sf}
\alias{SPAs_Scallop_sf}
\title{SPAs_Scallop_sf Metadata}
\format{
sf polygon
\describe{
\item{layer}{polygon identifier}
}
}
\usage{
SPAs_Scallop_sf
}
\description{
These are the inshore Scallop SPAs converted from polylines to polygons
and saved as sf objects.  Data was extracted from the scallop github site, and done
independently of the scallop unit.
}
\keyword{datasets}
