% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strata_Mar_sf.R
\docType{data}
\name{Strata_Mar_sf}
\alias{Strata_Mar_sf}
\title{Strata_Mar_sf Metadata}
\format{
sf multipolygon
\describe{
\item{StrataID}{polygon identifier}
}
}
\usage{
Strata_Mar_sf
}
\description{
These are the polygons from Maritimes strata, converted to an sf object from
the original "MaritimesRegionEcosystemAssessmentStrata(2014-).shp" shapefile.
}
\keyword{datasets}
