% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfClamFAs_sf.R
\docType{data}
\name{banks_sf}
\alias{banks_sf}
\title{SurfClamFAs_sf}
\format{
sf polygon
\describe{
\item{Name}{Name of the area}
}
}
\description{
An sf polygon for the existing Surf clam fishing areas.  The Grand Bank area is just NAFO areas 3LNO.
}
