% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banks_sf.R
\docType{data}
\name{banks_sf}
\alias{banks_sf}
\title{banks_sf}
\format{
sf polygon
\describe{
\item{Name}{Name of the Bank}
\item{SID}{Polygon Identifier}
}
}
\usage{
banks_sf
}
\description{
An sf polygon showing the areas of Banquereau Bank, Grand Banks, Sable Bank, Middle Bank,
Emerald Bank, LaHave Bank, Browns Bank, St Pierre Bank, Green Bank, and Burgeo Bank.  Each
polygon
}
\keyword{datasets}
